+++ src/Client.php
@@ -72,7 +72,6 @@ use RuntimeException;
  * @method Services\ApiTrackingService apiTracking()
  * @method Services\ShippingZoneService shippingZones()
  * @method Services\ApiKeyService apiKeys()
- * @method Services\CheckoutPageService checkoutPages()
  * @method Services\DisputeService disputes()
  * @method Services\FileService files()
  * @method Services\AttachmentService attachments()
@@ -82,6 +81,7 @@ use RuntimeException;
  * @method Services\WebhooksService webhooks()
  * @method Services\WebhookCredentialsService webhookCredentials()
  * @method Services\PlaidCredentialsService plaidCredentials()
+ * @method Services\ExperianCredentialsService experianCredentials()
  * @method Services\ValuesListService lists()
  * @method Services\ValuesListTrackingService listsTracking()
  * @method Services\WebhookTrackingService webhooksTracking()
@@ -99,9 +99,12 @@ final class Client
 
     public const SANDBOX_HOST = 'https://api-sandbox.rebilly.com';
 
+    /**
+     * @deprecated We don't use API version in URI anymore.
+     */
     public const CURRENT_VERSION = 'v2.1';
 
-    public const SDK_VERSION = '2.8.0';
+    public const SDK_VERSION = '2.9.0';
 
     private static $services = [
         'authenticationOptions' => Services\AuthenticationOptionsService::class,
@@ -137,7 +140,6 @@ final class Client
         'threeDSecure' => Services\ThreeDSecureService::class,
         'apiTracking' => Services\ApiTrackingService::class,
         'apiKeys' => Services\ApiKeyService::class,
-        'checkoutPages' => Services\CheckoutPageService::class,
         'disputes' => Services\DisputeService::class,
         'coupons' => Services\CouponService::class,
         'couponsRedemptions' => Services\RedemptionService::class,
@@ -154,6 +156,7 @@ final class Client
         'customerTimeline' => Services\CustomerTimelineService::class,
         'aml' => Services\AmlService::class,
         'plaidCredentials' => Services\PlaidCredentialsService::class,
+        'experianCredentials' => Services\ExperianCredentialsService::class,
         'paymentInstruments' => Services\PaymentInstrumentService::class,
     ];
 
@@ -272,7 +275,7 @@ final class Client
 
         // Prepare middleware stack
         $this->middleware = new Middleware\CompositeMiddleware(
-            new Middleware\BaseUri($this->createUri($baseUrl . '/' . self::CURRENT_VERSION)),
+            new Middleware\BaseUri($this->createUri($baseUrl)),
             $organizationId,
             new Middleware\UserAgent(self::SDK_VERSION),
             $authentication,